+++ service/hook.go
@@ -26,8 +26,11 @@ type CleanupFunc func() error
 // ConfigSetupFunc is a function that sets up a service configuration.
 type ConfigSetupFunc func(c config.Config) error
 
-// SetupFunc is a function that sets up a service.
-type SetupFunc func(workerGroup worker.Group) error
+// PreRunFunc is a function that runs before the workers are started.
+type PreRunFunc func() error
+
+// SetupWorkersFunc is a function that sets up workers for the service.
+type SetupWorkersFunc func(workerGroup worker.Group) error
 
 // SetupHTTPFunc is a function that sets up a service HTTP server.
 type SetupHTTPFunc func(server *http.Server) error
@@ -35,12 +38,14 @@ type SetupHTTPFunc func(server *http.Server) error
 type Hooks interface {
 	Cleanup(f CleanupFunc)
 	ConfigSetup(f ConfigSetupFunc)
-	Setup(f SetupFunc)
+	PreRun(f PreRunFunc)
+	SetupWorkers(f SetupWorkersFunc)
 	SetupHTTP(f SetupHTTPFunc)
 
 	invokeCleanup() error
 	invokeConfigSetup(c config.Config) error
-	invokeSetup(workerGroup worker.Group) error
+	invokePreRun() error
+	invokeSetupWorkers(workerGroup worker.Group) error
 	invokeSetupHTTP() (*http.Server, error)
 }
 
@@ -50,8 +55,10 @@ type hookstruct struct {
 	cleanup CleanupFunc
 	// configSetup is the setup configuration hook.
 	configSetup ConfigSetupFunc
-	// setup is the setup hook.
-	setup SetupFunc
+	// prerun is the prerun hook.
+	prerun PreRunFunc
+	// setupWorkers is the setupWorkers hook.
+	setupWorkers SetupWorkersFunc
 	// setupHTTP is the setup HTTP hook.
 	setupHTTP SetupHTTPFunc
 	// httpNewServer is a function that creates a new HTTP server. This is used
@@ -66,14 +73,19 @@ func (h *hookstruct) Cleanup(f CleanupFunc) {
 	h.cleanup = f
 }
 
-// SetupConfig registers a setup configuration hook.
+// ConfigSetup registers a configuration setup hook.
 func (h *hookstruct) ConfigSetup(f ConfigSetupFunc) {
 	h.configSetup = f
 }
 
-// Setup registers a setup hook.
-func (h *hookstruct) Setup(f SetupFunc) {
-	h.setup = f
+// PreRun registers a prerun hook.
+func (h *hookstruct) PreRun(f PreRunFunc) {
+	h.prerun = f
+}
+
+// SetupWorkers registers a worker setup hook.
+func (h *hookstruct) SetupWorkers(f SetupWorkersFunc) {
+	h.setupWorkers = f
 }
 
 // SetupHTTP registers a setup HTTP hook.
@@ -97,10 +109,18 @@ func (h *hookstruct) invokeConfigSetup(c config.Config) error {
 	return nil
 }
 
-// invokeSetup invokes the setup hook.
-func (h *hookstruct) invokeSetup(workerGroup worker.Group) error {
-	if h.setup != nil {
-		return h.setup(workerGroup)
+// invokePreRun invokes the prerun hook.
+func (h *hookstruct) invokePreRun() error {
+	if h.prerun != nil {
+		return h.prerun()
+	}
+	return nil
+}
+
+// invokeSetupWorkers invokes the setup hook.
+func (h *hookstruct) invokeSetupWorkers(workerGroup worker.Group) error {
+	if h.setupWorkers != nil {
+		return h.setupWorkers(workerGroup)
 	}
 	return nil
 }