+++ service/hook_test.go
@@ -30,7 +30,7 @@ func Test_hooks_NoHooks(t *testing.T) {
 	if err := h.invokeConfigSetup(nil); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
-	if err := h.invokeSetup(nil); err != nil {
+	if err := h.invokeSetupWorkers(nil); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if err := h.invokeCleanup(); err != nil {
@@ -72,20 +72,37 @@ func Test_hooks_InvokeConfigSetup(t *testing.T) {
 	}
 }
 
-// Test_hooks_InvokeSetup tests that the setup hook is invoked.
-func Test_hooks_InvokeSetup(t *testing.T) {
+// Test_hooks_InvokePreRun tests that the pre-run hook is invoked.
+func Test_hooks_InvokePreRun(t *testing.T) {
 	t.Parallel()
 	wasInvoked := false
 	h := &hookstruct{}
-	h.Setup(func(workerGroup worker.Group) error {
+	h.PreRun(func() error {
 		wasInvoked = true
 		return nil
 	})
-	if err := h.invokeSetup(nil); err != nil {
+	if err := h.invokePreRun(); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
 	if !wasInvoked {
-		t.Error("expected setup hook to be invoked")
+		t.Error("expected pre-run hook to be invoked")
+	}
+}
+
+// Test_hooks_InvokeSetupWorkers tests that the setupWorkers hook is invoked.
+func Test_hooks_InvokeSetupWorkers(t *testing.T) {
+	t.Parallel()
+	wasInvoked := false
+	h := &hookstruct{}
+	h.SetupWorkers(func(workerGroup worker.Group) error {
+		wasInvoked = true
+		return nil
+	})
+	if err := h.invokeSetupWorkers(nil); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+	if !wasInvoked {
+		t.Error("expected setupWorkers hook to be invoked")
 	}
 }